+++ internal/app/kubernetes/cluster.go
@@ -7,13 +7,10 @@ package kubernetes
 import (
 	"context"
 	"fmt"
-	"strings"
 
-	"github.com/clivern/beetle/internal/app/model"
 	"github.com/clivern/beetle/internal/app/module"
 
 	"github.com/spf13/viper"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/tools/clientcmd"
@@ -93,166 +90,4 @@ func (c *Cluster) Ping(ctx context.Context) (bool, error) {
 	}
 
 	return (string(data) != ""), nil
-}
-
-// GetNamespaces gets a list of cluster namespaces
-func (c *Cluster) GetNamespaces(ctx context.Context) ([]model.Namespace, error) {
-	result := []model.Namespace{}
-
-	fs := module.FileSystem{}
-
-	if !fs.FileExists(c.Kubeconfig) {
-		return result, fmt.Errorf(
-			"cluster [%s] config file [%s] not exist",
-			c.Name,
-			c.Kubeconfig,
-		)
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", c.Kubeconfig)
-
-	if err != nil {
-		return result, err
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-
-	if err != nil {
-		return result, err
-	}
-
-	data, err := clientset.CoreV1().Namespaces().List(ctx, metav1.ListOptions{})
-
-	if err != nil {
-		return result, err
-	}
-
-	for _, namespace := range data.Items {
-		result = append(result, model.Namespace{
-			Name:   namespace.ObjectMeta.Name,
-			UID:    string(namespace.ObjectMeta.UID),
-			Status: strings.ToLower(string(namespace.Status.Phase)),
-		})
-	}
-
-	return result, nil
-}
-
-// GetNamespace gets a namespace by name
-func (c *Cluster) GetNamespace(ctx context.Context, name string) (model.Namespace, error) {
-	result := model.Namespace{}
-
-	fs := module.FileSystem{}
-
-	if !fs.FileExists(c.Kubeconfig) {
-		return result, fmt.Errorf(
-			"cluster [%s] config file [%s] not exist",
-			c.Name,
-			c.Kubeconfig,
-		)
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", c.Kubeconfig)
-
-	if err != nil {
-		return result, err
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-
-	if err != nil {
-		return result, err
-	}
-
-	namespace, err := clientset.CoreV1().Namespaces().Get(ctx, name, metav1.GetOptions{})
-
-	if err != nil {
-		return result, err
-	}
-
-	result.Name = namespace.ObjectMeta.Name
-	result.UID = string(namespace.ObjectMeta.UID)
-	result.Status = strings.ToLower(string(namespace.Status.Phase))
-
-	return result, nil
-}
-
-// GetDeployments gets a list of deployments
-func (c *Cluster) GetDeployments(ctx context.Context, namespace string, labels string) ([]model.Deployment, error) {
-	result := []model.Deployment{}
-
-	fs := module.FileSystem{}
-
-	if !fs.FileExists(c.Kubeconfig) {
-		return result, fmt.Errorf(
-			"cluster [%s] config file [%s] not exist",
-			c.Name,
-			c.Kubeconfig,
-		)
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", c.Kubeconfig)
-
-	if err != nil {
-		return result, err
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-
-	if err != nil {
-		return result, err
-	}
-
-	data, err := clientset.AppsV1().Deployments(namespace).List(ctx, metav1.ListOptions{
-		LabelSelector: labels,
-	})
-
-	if err != nil {
-		return result, err
-	}
-
-	for _, deployment := range data.Items {
-		result = append(result, model.Deployment{
-			Name: deployment.ObjectMeta.Name,
-		})
-	}
-
-	return result, nil
-}
-
-// GetDeployment gets a deployment by name
-func (c *Cluster) GetDeployment(ctx context.Context, namespace, name string) (model.Deployment, error) {
-	result := model.Deployment{}
-
-	fs := module.FileSystem{}
-
-	if !fs.FileExists(c.Kubeconfig) {
-		return result, fmt.Errorf(
-			"cluster [%s] config file [%s] not exist",
-			c.Name,
-			c.Kubeconfig,
-		)
-	}
-
-	config, err := clientcmd.BuildConfigFromFlags("", c.Kubeconfig)
-
-	if err != nil {
-		return result, err
-	}
-
-	clientset, err := kubernetes.NewForConfig(config)
-
-	if err != nil {
-		return result, err
-	}
-
-	deployment, err := clientset.AppsV1().Deployments(namespace).Get(ctx, name, metav1.GetOptions{})
-
-	if err != nil {
-		return result, err
-	}
-
-	result.Name = deployment.ObjectMeta.Name
-
-	return result, nil
 }
