+++ internal/app/kubernetes/configmap.go
@@ -1,0 +1,10 @@
+// Copyright 2020 Clivern. All rights reserved.
+// Use of this source code is governed by the MIT
+// license that can be found in the LICENSE file.
+
+package kubernetes
+
+// GetConfigMap gets a configmap data
+func (c *Cluster) GetConfigMap() (string, error) {
+	return "", nil
+}
